refactor(apollo-client): move '@ApolloJsonValue' processor
