+++ apollo-client/src/main/java/com/ctrip/framework/apollo/spring/annotation/ApolloAnnotationProcessor.java
@@ -3,13 +3,23 @@ package com.ctrip.framework.apollo.spring.annotation;
 import com.ctrip.framework.apollo.Config;
 import com.ctrip.framework.apollo.ConfigChangeListener;
 import com.ctrip.framework.apollo.ConfigService;
+import com.ctrip.framework.apollo.build.ApolloInjector;
 import com.ctrip.framework.apollo.model.ConfigChangeEvent;
+import com.ctrip.framework.apollo.spring.property.PlaceholderHelper;
+import com.ctrip.framework.apollo.spring.property.SpringValue;
+import com.ctrip.framework.apollo.spring.property.SpringValueRegistry;
+import com.ctrip.framework.apollo.spring.util.SpringInjector;
+import com.ctrip.framework.apollo.util.ConfigUtil;
 import com.google.common.base.Preconditions;
+import com.google.gson.Gson;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.lang.reflect.Type;
 import java.util.Set;
 
 import com.google.common.collect.Sets;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
@@ -24,6 +34,19 @@ import org.springframework.util.ReflectionUtils;
  */
 public class ApolloAnnotationProcessor extends ApolloProcessor implements BeanFactoryAware {
 
+  private static final Logger logger = LoggerFactory.getLogger(ApolloAnnotationProcessor.class);
+  private static final Gson GSON = new Gson();
+
+  private final ConfigUtil configUtil;
+  private final PlaceholderHelper placeholderHelper;
+  private final SpringValueRegistry springValueRegistry;
+
+  public ApolloAnnotationProcessor() {
+    configUtil = ApolloInjector.getInstance(ConfigUtil.class);
+    placeholderHelper = SpringInjector.getInstance(PlaceholderHelper.class);
+    springValueRegistry = SpringInjector.getInstance(SpringValueRegistry.class);
+  }
+
   /**
    * resolve the expression.
    */
@@ -32,11 +55,13 @@ public class ApolloAnnotationProcessor extends ApolloProcessor implements BeanFa
   @Override
   protected void processField(Object bean, String beanName, Field field) {
     this.processApolloConfig(bean, field);
+    this.processApolloJsonValue(bean, beanName, field);
   }
 
   @Override
   protected void processMethod(final Object bean, String beanName, final Method method) {
     this.processApolloConfigChangeListener(bean, method);
+    this.processApolloJsonValue(bean, beanName, method);
   }
 
   private void processApolloConfig(Object bean, Field field) {
@@ -81,17 +106,97 @@ public class ApolloAnnotationProcessor extends ApolloProcessor implements BeanFa
       }
     };
 
-    Set<String> interestedKeys = annotatedInterestedKeys.length > 0 ? Sets.newHashSet(annotatedInterestedKeys) : null;
-    Set<String> interestedKeyPrefixes = annotatedInterestedKeyPrefixes.length > 0 ? Sets.newHashSet(annotatedInterestedKeyPrefixes) : null;
+    Set<String> interestedKeys =
+        annotatedInterestedKeys.length > 0 ? Sets.newHashSet(annotatedInterestedKeys) : null;
+    Set<String> interestedKeyPrefixes =
+        annotatedInterestedKeyPrefixes.length > 0 ? Sets.newHashSet(annotatedInterestedKeyPrefixes)
+            : null;
 
     for (String namespace : namespaces) {
-      Config config = ConfigService.getConfig(this.configurableBeanFactory.resolveEmbeddedValue(namespace));
+      Config config = ConfigService
+          .getConfig(this.configurableBeanFactory.resolveEmbeddedValue(namespace));
 
       if (interestedKeys == null && interestedKeyPrefixes == null) {
         config.addChangeListener(configChangeListener);
       } else {
         config.addChangeListener(configChangeListener, interestedKeys, interestedKeyPrefixes);
       }
+    }
+  }
+
+
+  private void processApolloJsonValue(Object bean, String beanName, Field field) {
+    ApolloJsonValue apolloJsonValue = AnnotationUtils.getAnnotation(field, ApolloJsonValue.class);
+    if (apolloJsonValue == null) {
+      return;
+    }
+    String placeholder = apolloJsonValue.value();
+    Object propertyValue = placeholderHelper
+        .resolvePropertyValue(this.configurableBeanFactory, beanName, placeholder);
+
+    // propertyValue will never be null, as @ApolloJsonValue will not allow that
+    if (!(propertyValue instanceof String)) {
+      return;
+    }
+
+    boolean accessible = field.isAccessible();
+    field.setAccessible(true);
+    ReflectionUtils
+        .setField(field, bean, parseJsonValue((String) propertyValue, field.getGenericType()));
+    field.setAccessible(accessible);
+
+    if (configUtil.isAutoUpdateInjectedSpringPropertiesEnabled()) {
+      Set<String> keys = placeholderHelper.extractPlaceholderKeys(placeholder);
+      for (String key : keys) {
+        SpringValue springValue = new SpringValue(key, placeholder, bean, beanName, field, true);
+        springValueRegistry.register(this.configurableBeanFactory, key, springValue);
+        logger.debug("Monitoring {}", springValue);
+      }
+    }
+  }
+
+  private void processApolloJsonValue(Object bean, String beanName, Method method) {
+    ApolloJsonValue apolloJsonValue = AnnotationUtils.getAnnotation(method, ApolloJsonValue.class);
+    if (apolloJsonValue == null) {
+      return;
+    }
+    String placeHolder = apolloJsonValue.value();
+
+    Object propertyValue = placeholderHelper
+        .resolvePropertyValue(this.configurableBeanFactory, beanName, placeHolder);
+
+    // propertyValue will never be null, as @ApolloJsonValue will not allow that
+    if (!(propertyValue instanceof String)) {
+      return;
+    }
+
+    Type[] types = method.getGenericParameterTypes();
+    Preconditions.checkArgument(types.length == 1,
+        "Ignore @Value setter {}.{}, expecting 1 parameter, actual {} parameters",
+        bean.getClass().getName(), method.getName(), method.getParameterTypes().length);
+
+    boolean accessible = method.isAccessible();
+    method.setAccessible(true);
+    ReflectionUtils.invokeMethod(method, bean, parseJsonValue((String) propertyValue, types[0]));
+    method.setAccessible(accessible);
+
+    if (configUtil.isAutoUpdateInjectedSpringPropertiesEnabled()) {
+      Set<String> keys = placeholderHelper.extractPlaceholderKeys(placeHolder);
+      for (String key : keys) {
+        SpringValue springValue = new SpringValue(key, apolloJsonValue.value(), bean, beanName,
+            method, true);
+        springValueRegistry.register(this.configurableBeanFactory, key, springValue);
+        logger.debug("Monitoring {}", springValue);
+      }
+    }
+  }
+
+  private Object parseJsonValue(String json, Type targetType) {
+    try {
+      return GSON.fromJson(json, targetType);
+    } catch (Throwable ex) {
+      logger.error("Parsing json '{}' to type {} failed!", json, targetType, ex);
+      throw ex;
     }
   }
